NewPlayerImpl: use a single instance internally, always initted
This is step 1 towards having a better separation between the init
states of the internal ExoPlayer object; we want to make sure that
most functionality is only available when the player is initialized,
instead of returning arbitrary values (or crashing).

On its own, this is probably buggy, because it will call the ExoPlayer
object after `release()` is called if any of the by-properties are
accessed.

See the next commits.