fix: don’t cast LocalContext to Activity
This only works if the surrounding Context is from an Activity, but
when embedding in NewPipe proper, we are in a Fragment. So the cast
will fail.

Instead, let’s use the util function we already had that iterates
through the `ContextWrapper`s until it finds an activity, and use that
everywhere.