fix: don’t cast LocalContext to Activity
This only works if the surrounding Context is from an Activity, but
when embedding in NewPipe proper, we are in a Fragment. So the cast
will fail.

Instead, for now we inject the Activity in the model. But what we
should actually do is forward the data we need from outside.